+++ backend/Application/Api/GraphQL/Versions/Versions.cs
@@ -1,0 +1,4 @@
+public class Versions
+{
+    public String BackendVersion { get; set; }
+}